Channels are displayed on sidebar
Channels component gets channels on mount and adds them to state.
The component use the new displayChannels method to map over the array of channels and create a menu item for each one.
Currently clicking a channel only console logs it, so the next step is to add selected channel to global state.